% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymolar.R
\name{loadPackages}
\alias{loadPackages}
\title{This function checks if required Packages are available or not. If there are any packages are missing
then it creats a shiny app with message of missing apps. Otherwise it returns NULL object.}
\usage{
loadPackages(requiredPackages = c("shiny"))
}
\arguments{
\item{requiredPackages}{}
}
\value{
Shiny app object
}
\description{
This function checks if required Packages are available or not. If there are any packages are missing
then it creats a shiny app with message of missing apps. Otherwise it returns NULL object.
}
